Organized image processing tests and added access control tests for image generation.
