+++ tests/cypress/integration/image-processing/pdf-read.test.js
@@ -1,6 +1,6 @@
 /* eslint jest/expect-expect: 0 */
 
-import { getPDFData } from '../plugins/functions';
+import { getPDFData } from '../../plugins/functions';
 
 describe('PDF read Tests', () => {
